# swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Alkemy Wallet",
    "version": "v1"
  },
  "paths": {
...
    "/Roles": {
      "get": {
        "tags": [
          "Roles"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/RolesDTO"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RolesDTO"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/RolesDTO"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Roles"
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "required": [
                  "Description",
                  "Name"
                ],
                "type": "object",
                "properties": {
                  "Name": {
                    "type": "string"
                  },
                  "Description": {
                    "type": "string"
                  }
                }
              },
              "encoding": {
                "Name": {
                  "style": "form"
                },
                "Description": {
                  "style": "form"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/Roles/{id}": {
      "get": {
        "tags": [
          "Roles"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "put": {
        "tags": [
          "Roles"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "Name": {
                    "maxLength": 20,
                    "minLength": 2,
                    "type": "string"
                  },
                  "Description": {
                    "maxLength": 50,
                    "minLength": 2,
                    "type": "string"
                  }
                }
              },
              "encoding": {
                "Name": {
                  "style": "form"
                },
                "Description": {
                  "style": "form"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "delete": {
        "tags": [
          "Roles"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Role"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Role"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Role"
                }
              }
            }
          }
        }
      }
    },
...
    "/Users": {
      "get": {
        "tags": [
          "Users"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/UserDTO"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserDTO"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserDTO"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Users"
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "required": [
                  "Email",
                  "First_name",
                  "Last_name",
                  "Password",
                  "Points"
                ],
                "type": "object",
                "properties": {
                  "First_name": {
                    "type": "string"
                  },
                  "Last_name": {
                    "type": "string"
                  },
                  "Email": {
                    "type": "string"
                  },
                  "Password": {
                    "maxLength": 20,
                    "type": "string"
                  },
                  "Points": {
                    "type": "integer",
                    "format": "int32"
                  }
                }
              },
              "encoding": {
                "First_name": {
                  "style": "form"
                },
                "Last_name": {
                  "style": "form"
                },
                "Email": {
                  "style": "form"
                },
                "Password": {
                  "style": "form"
                },
                "Points": {
                  "style": "form"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/Users/{id}": {
      "get": {
        "tags": [
          "Users"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "put": {
        "tags": [
          "Users"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "First_name": {
                    "maxLength": 100,
                    "minLength": 2,
                    "type": "string"
                  },
                  "Last_name": {
                    "maxLength": 100,
                    "minLength": 2,
                    "type": "string"
                  },
                  "Password": {
                    "type": "string"
                  },
                  "Points": {
                    "type": "integer",
                    "format": "int32"
                  },
                  "Rol_id": {
                    "type": "integer",
                    "format": "int32"
                  }
                }
              },
              "encoding": {
                "First_name": {
                  "style": "form"
                },
                "Last_name": {
                  "style": "form"
                },
                "Password": {
                  "style": "form"
                },
                "Points": {
                  "style": "form"
                },
                "Rol_id": {
                  "style": "form"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "delete": {
        "tags": [
          "Users"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Account": {
        "required": [
          "creationDate",
          "isBlocked",
          "money",
          "user_id"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "creationDate": {
            "type": "string",
            "format": "date-time"
          },
          "money": {
            "maximum": 9999999999999,
            "minimum": 0,
            "type": "number",
            "format": "float"
          },
          "isBlocked": {
            "type": "boolean"
          },
          "user_id": {
            "type": "integer",
            "format": "int32"
          },
          "user": {
            "$ref": "#/components/schemas/User"
          },
          "fixedTermDeposit": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/FixedTermDeposit"
            },
            "nullable": true
          },
          "transaction": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Transaction"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "FixedTermDeposit": {
        "required": [
          "account_id",
          "amount",
          "closing_date",
          "creation_date",
          "user_id"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "user_id": {
            "type": "integer",
            "format": "int32"
          },
          "user": {
            "$ref": "#/components/schemas/User"
          },
          "account_id": {
            "type": "integer",
            "format": "int32"
          },
          "account": {
            "$ref": "#/components/schemas/Account"
          },
          "amount": {
            "maxLength": 255,
            "type": "number",
            "format": "float"
          },
          "creation_date": {
            "type": "string",
            "format": "date-time"
          },
          "closing_date": {
            "type": "string",
            "format": "date-time"
          }
        },
        "additionalProperties": false
      },
      "Role": {
        "required": [
          "description",
          "name"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "name": {
            "maxLength": 100,
            "minLength": 1,
            "type": "string"
          },
          "description": {
            "maxLength": 100,
            "minLength": 1,
            "type": "string"
          },
          "user": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/User"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "RolesDTO": {
        "required": [
          "description",
          "name"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "name": {
            "minLength": 1,
            "type": "string"
          },
          "description": {
            "minLength": 1,
            "type": "string"
          }
        },
        "additionalProperties": false
      },
      "Transaction": {
        "required": [
          "amount",
          "concept",
          "date",
          "to_Account",
          "type",
          "user_id"
        ],
        "type": "object",
        "properties": {
          "transaction_id": {
            "type": "integer",
            "format": "int32"
          },
          "amount": {
            "maxLength": 100,
            "type": "integer",
            "format": "int32"
          },
          "concept": {
            "maxLength": 100,
            "minLength": 1,
            "type": "string"
          },
          "date": {
            "type": "string",
            "format": "date-time"
          },
          "type": {
            "maxLength": 100,
            "minLength": 1,
            "type": "string"
          },
          "user_id": {
            "type": "integer",
            "format": "int32"
          },
          "user": {
            "$ref": "#/components/schemas/User"
          },
          "to_Account": {
            "type": "integer",
            "format": "int32"
          },
          "account_id": {
            "type": "integer",
            "format": "int32"
          },
          "account": {
            "$ref": "#/components/schemas/Account"
          }
        },
        "additionalProperties": false
      },
      "User": {
        "required": [
          "email",
          "first_name",
          "last_name",
          "password",
          "points",
          "rol_id"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "first_name": {
            "maxLength": 100,
            "minLength": 1,
            "type": "string"
          },
          "last_name": {
            "maxLength": 100,
            "minLength": 1,
            "type": "string"
          },
          "email": {
            "maxLength": 150,
            "minLength": 7,
            "type": "string"
          },
          "password": {
            "maxLength": 70,
            "minLength": 4,
            "type": "string"
          },
          "points": {
            "maxLength": 50,
            "type": "integer",
            "format": "int32"
          },
          "rol_id": {
            "maxLength": 100,
            "type": "integer",
            "format": "int32"
          },
          "role": {
            "$ref": "#/components/schemas/Role"
          },
          "account": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Account"
            },
            "nullable": true
          },
          "transaction": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Transaction"
            },
            "nullable": true
          },
          "fixedTermDeposit": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/FixedTermDeposit"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "UserDTO": {
        "required": [
          "email",
          "first_name",
          "last_name",
          "rol_id"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "first_name": {
            "minLength": 1,
            "type": "string"
          },
          "last_name": {
            "minLength": 1,
            "type": "string"
          },
          "email": {
            "minLength": 1,
            "type": "string"
          },
          "rol_id": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      }
    },
    "securitySchemes": {
      "Bearer": {
        "type": "apiKey",
        "description": "Jwt Authorization",
        "name": "Authorization",
        "in": "header"
      }
    }
  },
  "security": [
    {
      "Bearer": [ ]
    }
  ]
}